Struct probor::Decoder

pub struct Decoder<R> { /* fields omitted */ }

The actual decoder type definition

Methods

impl<R> Decoder<R> where
R: ReadBytesExt,

pub fn new(c: Config, r: R) -> Decoder<R>

pub fn into_reader(self) -> R

pub fn kernel(&mut self) -> &mut Kernel<R>

pub fn simple(&mut self) -> Result<Simple, DecodeError>

pub fn bool(&mut self) -> Result<bool, DecodeError>

pub fn u8(&mut self) -> Result<u8, DecodeError>

pub fn u16(&mut self) -> Result<u16, DecodeError>

pub fn u32(&mut self) -> Result<u32, DecodeError>

pub fn u64(&mut self) -> Result<u64, DecodeError>

pub fn i8(&mut self) -> Result<i8, DecodeError>

pub fn i16(&mut self) -> Result<i16, DecodeError>

pub fn i32(&mut self) -> Result<i32, DecodeError>

pub fn i64(&mut self) -> Result<i64, DecodeError>

pub fn int(&mut self) -> Result<Int, DecodeError>

pub fn f16(&mut self) -> Result<f32, DecodeError>

pub fn f32(&mut self) -> Result<f32, DecodeError>

pub fn f64(&mut self) -> Result<f64, DecodeError>

pub fn read_bytes(&mut self, b: &mut [u8]) -> Result<usize, DecodeError>

Decode a single byte string into the given buffer.

The provided buffer must be large enough to hold the entire byte string, otherwise an error is returned.

Please note that indefinite byte strings are not supported by this method (Consider using Decoder::bytes_iter() for this use-case).

pub fn bytes(&mut self) -> Result<Vec<u8>, DecodeError>

Decode a single byte string.

Please note that indefinite byte strings are not supported by this method (Consider using Decoder::bytes_iter() for this use-case).

pub fn bytes_iter(&mut self) -> Result<BytesIter<R>, DecodeError>

Decode an indefinite byte string.

pub fn text(&mut self) -> Result<String, DecodeError>

Decode a single UTF-8 encoded String.

Please note that indefinite strings are not supported by this method (Consider using Decoder::text_iter() for this use-case).

pub fn text_iter(&mut self) -> Result<TextIter<R>, DecodeError>

Decode an indefinite string.

pub fn tag(&mut self) -> Result<Tag, DecodeError>

Decode a Tag. If no tag is found an UnexpectedType error is returned.

pub fn array(&mut self) -> Result<usize, DecodeError>

Decode the begin of an array. The length is returned unless it exceeds the configured maximum.

Please note that indefinite arrays are not supported by this method (Consider using Decoder::array_begin() for this use-case).

pub fn array_begin(&mut self) -> Result<(), DecodeError>

Decode the begin of an indefinite array. After this one can continue decoding items, but a Break value will be encountered at some unknown point.

(Consider using or_break around every decoding step within an indefinite array to handle this case).

pub fn object(&mut self) -> Result<usize, DecodeError>

Decode the begin of an object. The size (number of key-value pairs) is returned unless it exceeds the configured maximum.

Please note that indefinite objects are not supported by this method (Consider using Decoder::object_begin for this use-case).

pub fn object_begin(&mut self) -> Result<(), DecodeError>

Decode the begin of an indefinite object. After this one can continue decoding items, but a Break value will be encountered at some unknown point.

(Consider using or_break around every decoding step within an indefinite object to handle this case).

impl<R> Decoder<R> where
R: Skip + ReadBytesExt,

pub fn skip(&mut self) -> Result<(), DecodeError>

Skip over a single CBOR value.

Please note that this function does not validate the value hence it might not even be well-formed CBOR. Instead skip is an optimisation over GenericDecoder::value() and generally only determines as much information as necessary to safely skip a value without keeping all of it in memory.

impl<R> Decoder<R> where
R: ReadSlice + ReadBytesExt,

pub fn text_borrow(&mut self) -> Result<&str, DecodeError>

Decode a single UTF-8 encoded String and borrow it from underlying buffer instead of allocating.

Please note that indefinite strings are not supported by this method.

pub fn bytes_borrow(&mut self) -> Result<&[u8], DecodeError>

Decode a single byte string and borrow it from underlying buffer instead of allocating.

Please note that indefinite byte strings are not supported by this method.